Move query parsing to Query class
